Move xml to array conversion fn to util
